Improve SignOutReviewWindow
Make level of learning hidden, alongside topic list, if 135 is not checked. Made visible when 135 is checked.
Make labels for level of learning and tutor dropdowns.
Enforce that if 135 is checked, level of learning cannot be null, and at least 1 selection must be made for topics. If 135 is not checked, only tutor must be selected.
Since level of learning can be null, it would not be added to the DB. Added identical functionality if topics is null.